Always fully build plugins with the main program (#252)
This commit solves a situation when the sidecar  fails to load
plugins with an error: "plugin was built with a different
version of package ...".  It appears that some build artifacts were
stored in docker cache and reused in future builds, causing this
incompatibility. The `--no-cache` flag added to the `docker build`
command is fixing the issue.