Allow duplicate keys in ptpconfig
When initially designing PTP config parsing, there was an assumption
that for each section, all of the keys were unique.  Based on this
assumption, key:value pairs for a given section were stored as a
map[string][string].

With the introduction of unicast_master_table, this assumption is no
longer true, as in the case of the UDPv4 key, there can be multiple
entries with the same key.  As such, the config data structure is
changed to maintain a list of individual key:value pairs, rather than a
map of keys to values.

In addition, some restructuring was done for places that relied on the
pre-existing structure, moving most of that functionality to utility
functions in config.go